Add scaffold proto
`proto.dali` is a DaLanguage program written in the instruction layer
with the sole purpose of reducing DaLanguage programs written in the
meta layer to programs in the instruction layer. Bootstrapping is the
main use of `proto.dali`.

We'll refer to `proto.dali` as "proto" from here on out.

While proto will follow the parsing transformation rules of the meta
layer, it will not implement any of the semantics of the trait layer.
The information in the trait layer will be discarded to avoid adding
complexity of proto.

The bootstrapper is a program external to DaLanguage. The only
requirement for a bootstrapper is that it can execute arbitrary
instruction layer programs and that is has the initial values necessary
to execute proto and the self-hosted compiler. DaLanguage will hold
multiple bootstrappers, but only one may be used during the
bootstrapping process based on need and platform.

And outline of the bootstrapping process is as follows:

1. A bootstrapper is used to execute proto.
2. Proto is used to reduce the self-hosted compiler's source files to
   the instruction system. The result of this reduction shall be
   referred to as the reduced compiler.
3. The bootstrapper is used to execute the reduced compiler.
4. The reduced compiler is used to compile the DaLanguage sources of the
   self-hosted compiler, creating the final compiler in a form native to
   the current platform. This ends the bootstrapping process.

Note: this is an early draft of the bootstrapping process, it is subject
      to change in the future.

The current version of proto is a placeholder to build off of. Currently
it only echos out its input.